requirements.txt
moviepy==1.0.2
pillow
numpy
whisper
google-auth
google-auth-oauthlib
requests
python-dotenv
